Add sub_test filtering for Jira 274
Jira 274 (Master got an xSocket) is a sporadic failure we see under gasnet-fast
for arrays/johnk/associative/tooFewKeys and a few other tests.

This adds sub_test filtering so it's easier to pick out when this error errors.
It doesn't happen too often in nightly testing, but I'm doing a bunch of trials
for the "qthreads use chapel allocator" work, and it comes up often enough to
warrant filtering.