Allow fields and vars to be 'void', make range.stride void if not stridable
Remove the code that issues the error "illegal use of function that does not
return a value".  That is legal now as long as the 'void' result is only
assigned to a 'void' variable.

Allow '_void' to be returned from functions as a param.

At the end of function resolution, clean up by removing 'void' variables and
fields.  Remove PRIM_DEREF, PRIM_GET_MEMBER, PRIM_SET_MEMBER, etc. involving
'void'. Remove 'void' actuals and formals.  Change
'PRIM_MOVE(voidVar, aFunction())' into just a call to 'aFunction()'.

Add assignment operators that issue a compilerError for 'void = val' and
'val = void' assignments. Add a 'void = void' assignment that does nothing.

Add a blank 'void.writeThis(f)' method and make 'void' not be an 'isIoField' in
order to make it not print anything in writeln() and similar functions.

Add a compilerError in 'buildArray' for arrays of 'void'.

Add 'void' to the list of types resolveIntents can handle.

Allow 'autoCopy' and 'initCopy' for 'void' to pass through
'removeUnnecessaryAutoCopyCalls' even though they don't have an argument
anymore.

Removed .future files for tests of 'void' variables and fields.  Update two of
the 'void' test .good files to reflect current expected behavior.

As a bigger test of void fields, changed range._stride a void field if the
range is not stridable.  This required some cleanup like using a 'stride'
accessor function in several places instead of accessing the '_stride' field
directly.