Implement first cut at NUMA-localized NIC-registered memory.
When we're using mem=jemalloc and a tasking layer like tasks=qthreads
with a fixed number of pthreads, each pthread has an arena permanently
assigned to it.  Here we take advantage of that to NUMA-localize any
NIC-registered memory.  Once we've gotten the desired shared heap from
the comm layer in mem=jemalloc, we block-localize its subchunks across
the (presumably NUMA) sublocales and then manage each subchunk as a
separate arena.  The arena extension logic always gets more memory from
the sublocale the thread is running on, so a task that allocates memory
will always get it from the sublocale its host thread is bound to.  In
DefaultRectangular we allocate multi-ddata array chunks from the bodies
of on-stmts that go to the NUMA sublocales, so those array chunks will
end up properly localized without us having to touch or migrate them.
And since jemalloc internally arranges to free memory back to the same
arena from which it was allocated, we don't have to do anything special
for free().

Most of the work here is in mem-jemalloc and most of that is to support
multiple heaps (at most 4, currently) instead of just 1.  Once we know
we have a desired shared from the comm layer, if we have multiple NUMA
domains (it's too early to ask the tasking layer about this so we ask
the topology layer), call the topology layer to block-localize that
heap, and then manage the localized blocks as our arenas.  (For reasons
that are not entirely clear at this point we still have to touch the
pages of these blocks while running on the appropriate NUMA domains to
get the actual memory localized, so that is also done at this time.
We'll look further into why this is needed later.)

In the topology layer, there are new functions to get and set execution
locality.  And, what used to be a single function to localize either the
whole of an area of memory of its subchunks has been turned into two
functions that do specifically those tasks, with common code factored
out into helpers.  There are also some changes for the sake of clarity,
such as moving flag settings closer to where they are used.

At the top surface of the memory layer in chpl-mem.h, we now localize
memory only if the allocator does not do so during allocation and the
allocation is large enough (a total of at least 2 pages across all the
sublocales) that it seems like it would help.  For now, jemalloc is the
only memory layer that can allocate localized memory directly, and even
it can only do so when the comm layer provides the desired shared heap.

Finally, while I was here, I also removed a superfluous blank line from
the array-loc test.