Add parens around generated ternary ops
Unless --no-denormalize was thrown, we were seeing issues with
test/errhandling/catch-match.chpl. The issue was that denormalize would
merge a dynamic cast ternary expression into a conditional if it was used
only once. That would be OK, except that the ternary operator has lower
precedence in C than comparison, which led to a type mismatch from the C
compiler.

The fix here is simple: simply generate parentheses around the ternary
operator in the generated C code.