Optimize serial iteration over ArrayViews with chpl__serialViewIter
DefaultRectangularArr used to be able to optimize serial iteration by
avoiding calls to dsiAccess and doing the appropriate indexing math
inside the serial iterator. When ArrayViews landed this was no longer
possible for slices, rank-changes, and reindexes. While this didn't
present a problem for the flat locale-model, we observed regressions for
NUMA.

This commit implements chpl__serialViewIter so that domain-maps can
overload the iterator and potentially implement more efficient
iteration.

The DefaultRectangularArr overloads are capable of accepting an
ArrayView and utilizing its indexCache to support optimal iteration over
ArrayViews. In support of this, 'getRADDataIndex' was renamed to
'getDataIndex'.