Compute heap page size as and when needed; accept libhugetlbfs setting.
There are two primary changes here.  The first is that when computing
the heap page size, if we find the HUGETLB_DEFAULT_PAGE_SIZE environment
variable set then we simply accept that as the page size, rather than
doing any more work to see if a larger size seems to work.  This saves a
bit of time on Cray X* systems but much more importantly, it avoids the
page touching needed in order to explore for larger page sizes.  That
page touching could cause undesirable NUMA localization of some of the
memory in comm layer desired shared heaps.

The second change is to compute the heap page size the first time it is
requested, rather than doing it in an explicit initialization step ahead
of time.  This relieves us needing to move the init call around to make
sure it stays ahead of the first call to get the heap size it computed
as the runtime evolves.  (This was going to be an issue with an upcoming
change to NUMA-localize comm layer desired shared heaps.)  As a side
effect of this change I reworked how we decide whether to bring in the
pthreads.h #include file.