Add specialized overload for c_ptrTo for DefaultRectangular
ArrayViews over a DefaultRectangular array store enough information in
their indexCache to determine whether the array can be passed as a
single _ddata. The logic for getting that pointer remains the same, but
the runtime check will utilize the indexCache when possible.

Also fixes a compile-timer error in the general c_ptrTo([]) case for
ArrayViews over other distributions (e.g., BlockDist).