Update .good files for "illegal use of function that does not return a value"
These tests assign non-returning function calls to variables. They previously
got an error "illegal use of function that does not return a value", but now
cause the variable to have type 'void'.  In a few cases, this allowed the test
to compile and run, in a few it lead to other reasonable errors.

Update the .good files for the new behavior.  Add a prediff to remove a
long list of candidates for an unresolved call to '<<'. Future-ize a test
that tries to cast to a void value, not type.  It should issue an error, but
fails similarly to the futures in expressions/vass/cast-to-non-type-*.future.