Only NUMA-localize the desired shared heap for comm=ugni.
Heap localization isn't working yet for comm=gasnet, partly due to a
problem in the topology layer and partly due to a problem in the
mem=jemalloc page-touching workaround for the chpl-topo problem.  So,
just do desired shared heap localization for comm=ugni, where it does
work.

While here, I also fixed a minor potential problem in the topology
layer, where the getMemLocality function would say that an address was
bound to a certain NUMA domain even if that was only the first of more
than one it could be bound to.  Now we return "any" if the page could be
bound to more than one NUMA domain.