In normalize, global-ize call_tmps used to initialize refs too
This fixes a problem with refs to arrays in a global context.
For example:

  module bar {
    var dom = {1..10, 1..10};
    var A : [dom] int;

    ref R = A[1..4, 1..4]; // call_tmp = slice
    // slice freed here
  }

  use bar;

  proc main() {
    writeln("Setting A[1,1] to 1");
    A[1,1] = 1;
    writeln("R[1,1] is now     ", R[1,1]);

    for r in R do write(r);
    writeln();
  }

The problem is that the call_tmp storing A[1..4, 1..4]
is destroyed at the end of the module bar's init function.

This PR resolves the issue by migrating such call_tmps to global variables. It
adjusts moveGlobalDeclarationsToModuleScope in normalize to do so.

If the for_SymbolSymExprs loop here is costly, we could change normalize to
mark the call_tmps initializing user reference variables and then simply detect
the marked call_tmps in moveGlobalDeclarationsToModuleScope.