Changes reflecting review feedback.
Convert chpl_mem_localizationThreshold from a variable to a function.

Slightly rename chpl_mem_allocLocalizes() (to ...alloc_localizes()) and
recode to try to clarify how it returns an explicit answer from the
implementation if that provides one, or a default "no" answer.

In the mem=jemalloc layer, create several helper functions out of code
that was inline before, remove the compile-time limit on the number of
heaps, and add a constraint that if locModel==flat the allocator does
not localize.