class Error, catch blocks, and strict mode (#5655)
lowers catch:
a control flow structure that handles errors when paired with try or try!
  - catch (catches all errors, without access to the error)
  - catch err (catches all errors, with named access to the error)
  - catch err: ErrType (catches errors of a given subtype, with named access)

While implementing this feature, it was found that codegen improperly generated dynamic casts because ternary operators do not have sufficient precedence. @mppf fixed this by adding those parentheses during codegen.

Note that errors are not automatically deallocated when consumed by a catch.
  - This will be in a future PR and is tracked in #4977.

implements Error:
module ChapelError has been included in ChapelStandard. It contains class Error, which the error handling system uses as its base class. This is done through dtError, which is used in passes/errorHandling.cpp.

implements strict mode:
When the --strict-errors flag is passed to the compiler, all throwing calls must be marked with try or try!. This is implemented by a simple compiler flag that is used in passes/errorHandling.cpp.

additional changes:
- Updated parser for catch syntax
- Added CatchStmt
- Update TryStmt to include CatchStmts
- De-futurized and reorganized error handling tests
- Added tests for error cases