Add tests for returning various things
Unfortunately, this pattern is not currently working correctly:

  proc f(ref arg) return arg;

it should generate an autoCopy but does not.

Added some tests exploring this space a little bit.

These fail to add the required autoCopy:

  proc f(ref arg) return arg;
  proc f(const ref arg) return arg;

This adds the autoCopy correctly:

  proc f(arg) return arg;

The new tests return-global and return-in-arg test other cases that are
currently working but that I didn't feel had good enough coverage.