Enable initCopy to have a ref formal, use autoCopy for Owned
This is useful for Owned.

in resolveAutoCopyEtc, check for FLAG_COPY_MUTATES and use chpl_autoCopy
in that case (even for a user record; previously we would just use
chpl_initCopy there). The reason is that if FLAG_COPY_MUTATES is set, the
compiler isn't currently set up to properly const-check the late
additions of chpl_autoCopy.

in lvalueCheck, INTENT_IN/INTENT_CONST_IN will modify actual
and so should generate lvalue errors.

In resolveArgIntent, if the type has the flag FLAG_COPY_MUTATES, use
INTENT_REF after resolution; otherwise it can be INTENT_CONST_REF.