Add regexdna-redux.chpl to the release and submitted directories.
The Computer Language Benchmarks Game has recently undergone some changes with
regards to the regexdna benchmark, resulting in the updating and migrating of
many implementations to the new form.  This saves a copy of the updated version
to our submitted and release directories, so that we can track that in addition
to our original version.

Note that the new benchmark has slightly different expected output than the
original, leading to differences in the .good and .perfkeys files.  Other than
the obvious update to the .chpl file, all other support files were copied
verbatim from the original.