Improve rank-change DomainView performance
Calling ``downIdxToUpIdx``, or some similar translation function,
on each yielded indice is costly. Since the ``updom`` is a
DefaultRectangular, we can simply iterate over that instead
if ``downdom`` is a DefaultRectangularDom (or a view over one).
Inlining the ``downIdxToUpIdx`` function does help a little
bit though.

Adds the following helper functions/methods (similar to the
ArrayView variants):
- chpl__isDROrDRView
- baseDom.isSliceDomainView
- baseDom.isRankChangeDomainView
- baseDom.isReindexDomainView
- _getActualDomain
- chpl__isDomainView
- chpl__getActualDomain
- BaseDom.isDefaultRectangular