Fix a couple of corner cases
Now that my getRankChangeUpRanges() processes scalar arguments, I was
no longer triggering the error message we used to generate for
all-scalar arguments.  Here, I added an explicit varargs this()
function that takes all integral arguments to generate that error
instead.

I'd also accidentally broken the logic that makes the updomain an
empty domain in the event that one of the rank-changing indices is
out-of-bounds for the original domain (an empty intersection).
Restored that behavior, albeit in a different way.