Make type/param versions return themselves again
Rationale: They're used in where clauses and primitive expressions
  where their values need to be known.

This passes testing, but it's not clear that it should.  I.e., is
the 'return true' in the main case going to cause problems down
the line?  In what other places are statementLevel symbols
inserted where the return value is used?