Fix chpl__unref for ArrayViewRankChange
chpl__unref, given a rank-changed array, will now return an array
declared over a rank-changed domain. This violates some compiler logic
that inserts the chpl__unref. If the types were the same, then the
chpl__unref would be removed.

Instead, remove the chpl__unref implementation for non-view arrays and
instead of checking the types to see if we're dealing with an ArrayView,
check if the chpl__unref call will resolve at all.