Fix leak in ArrayViewRankChange
dsiBuildArray on a rank-changed domain would create a new array, but
never free'd it because the dsiDestroy signature was incorrect (it
accepts a boolean argument).

Even then, we don't want to always free the inner array instance.
Instead, add a boolean to the class to indicate when it should free the
array instance.

TODO: can this boolean be param?