Refactor dsiBuildNewRectangularDom()
Previously, when we supported a dsi call to build new domains from
domains, it took the indices in at the same time, so could avoid a
duplicated setup() call for distributed domains, which cut down on
communication.  When I eliminated that call and switched to a
dsiNewRectangularDom() on distributions followed by a dsiSetIndices(),
I was getting two setup calls and therefore increasing communication.
Here, I add an additional 'inds' argument to dsiNewRectangularDom()
which permits the indices to be set up the first time around.  I also
created two versions of the newRectangularDom() call, one which takes
indices, and one which doesn't; and the one that doesn't just passes
in an empty index set, which should be semantically equivalent to what
we had before.  My goal in doing this was to create just one dsi
routine to build rectangular domains, but two buildRectangularDom()
calls...

I still need to convert more of the old pairs of calls into the new
single call form.