Update debug output to reflect new earlier setting of dsiSetIndices
This raises a question as to whether using dsiSetIndices earlier in
this manner when we don't have indices to set adds unnecessary
overhead relative to the old code path.  I'm guessing that for
distributed domains it doesn't and am hoping it's in the noise for
local ones.  If not, we can either create two overloads of
dsiNewRectangularDom() or we can have it take a bool saying whether
or not the indices are useful (or check for the empty set).