Attempt to fix locality queries for default rectangular stuff
This is a failed five-minute fix (TM) to file and fix a test that
tries to use the array/domain locality queries for a default-rectangular
domain and array from different locales.  For some reason, the responses
for 'B' are not as expected, but I'm timing out for now.  Filing this
in lieu of an issue for now.