Build amudprun for the host (#5750)
Build amudprun for the host (#5750)

[reviewed by @ronawho ]

By default GASNet builds the amudprun launcher for the target machine (compute nodes). It should be built for the host (login node). However, the library on which its protocol is based is still needed for both the host and the target.

This change leaves the GASNet build untouched, allowing it to build the target version of the library, and adds a second build of amudprun, which runs on the host. This second build will only occur if CHPL_LAUNCHER=amudprun .

Tested on 64-bit ARM, Mac, Linux, and Cray XC.