Don't ever report a number of CPUs less than 1 (one).
A user issue (https://github.com/chapel-lang/chapel/issues/5536) reports
test failures that ultimately are caused by chpl_getNumPhysicalCpus()
returning 0 (zero).  This change ensures that both of the functions that
return CPU counts always return 1 (one) or higher.

In the next release much of this code should be reworked in favor of
getting the value from the topology layer, but that's too big a change
for code freeze week.