Handle .bad mismatch for configHelp.chpl
.bad was no longer matching after #5744.

Instead of adjusting .bad, I am adding a .prediff
that greps out most of the normal help message.

That way comparing against .bad will be insensitive to most of the
specifics of the help message that may again change in the future.

I am preventing the .prediff from messing with the output
if it already matches .good, just because I can.