Fix nolocal regression for printFinalMemStat
Background:

The amount of temporarily-used memory changed due to #5744
causing regression in test memory/shannon/printFinalMemStat

The test itself intentionally leaks 256 bytes.  This is probably
desirable to make sure that the stats we get using the
printFinalMemStat flag are correct.  The downfall of the test is
the changing amount of extra allocations/leaks introduced by
various configurations. These changes have been unpleasantly
frequent and have been turned up only after nightly testing,
causing testing noise and maintenance burden of updating .good.

Solutions we considered:

(a) Add a prediff to convert numbers to nnn. Still, leave unfiltered
    the number for "Current Allocated Memory" to ensure it remains 256.

(b) Add a .skipif that avoids testing this in any scenario where the
    result doesn't match the .good file.

(c) Free the buffers to avoid leaking.  Better yet, just create an array
    or something instead of manually calling the runtime function
    `chpl_mem_allocMany` to allocate memory.

Chosen solution:

We chose (a) because it is more airtight. Versus (c) it ensures
correct leaking amount. Versus (b) it avoids the danger of overlooking
breakage of --memStats on skipped configurations.