Add system config defines for amudprun when appropriate (#5780)
Add system config defines for amudprun when appropriate (#5780)

[reviewed by @gbtitus ]

In #5750 I left out two defines that were important for the case of running with GASNET_WORKERIP set. When amudprun is built for the target (the wrong machine), these defines are set by the configure script. I could not take advantage of the target configuration for the host, so in this change, the Makefile does a simple test to determine whether to set them.

Without these defines, a warning about lack of operating system capabilities appears, and was encountered by @benharsh .