Clean up readWriteThis for ranges
Commit 8842e921399019945e4603b156018257ab41315c introduced a modification
to the _alignment field even when writing a range. But writing a range
should be possible with a `const` range, so that's not a good strategy.

This commit instead computes a separate range variable with a normalized
alignment and runs isNaturallyAligned() on that. While there, I
applied the same logic to the cast from range to string, so that we
should not see the `align 0` be printed out there either.

The original commit was to make the test

  test/types/range/hilde/by.chpl

work properly without printing a lot of `align 0`s. It still works
after this commit.