Use const intent for atomic read/peek/waitFor/writeThis
A note about const-ness: these atomic types are records but
in order for record assigment to work, the read() functions
need to be able to work with a const RHS.

To enable that, the read/peek/waitFor/writeThis functions take in `this`
with const intent. That is reasonable even if the atomic is
implemented with a lock because the programmer can view it
as constant, and on good hardware it really will be. If we change
our mind about this, we will need to revisit the default `=`
generated for records containing atomic fields.