Add four tests of where clauses on initializers
The first case exercises using a where clause when the argument signature is
fully specified and otherwise identical.  The second case exercises using a
where clause when the argument signature of the two initializers is generic but
would otherwise lead to ambiguity during resolution.  The third case exercises
using a where clause when the arguments are fully specified and so the where
clause is redundant.  The fourth case exercises using a where clause to
eliminate other calls that the generic argument would otherwise allow (in this
case, reals).

Prior to my compiler modifications, three of these tests failed.