Allow initializers to accept where clauses
Prior to this change, we were failing to accept initializers that defined
where clauses because the class allocator did not recognize and copy the where
clause on the initializer with which it was paired.  This change copies the
where clause and ensures that it refers to the arguments in the _new method
rather than the arguments in the original initializer.