Don't mess up schedaffinity when touching desired heap for locality.
The code for touching desired heap pages into existence on the "right"
NUMA domains was careful to first record the thread's (process's) NUMA
locality and then restore it afterward.  Unfortunately, the function
used to get that locality was one designed to return the specific NUMA
domain associated with the thread and, as such, if the thread had
affinity to multiple domains the function reduced this to just the first
before it returned the value.  When we later "restored" that locality,
any NUMA domains the thread had originally had affinity to, other than
the first one, were lost.  The effect was observed in some runs on Cray
XE systems, where the reduced affinity had the effect of changing the
number of physical cores the program thought were available to it.

To work around this, move the page-touching localization code out of the
mem=jemalloc code and into the topology code, in a new function called
chpl_topo_touchMemFromSubloc().  Here we can save the original affinity
and restore it afterward without losing information.