Fix misc errors for distributed reindex
Passing the distribution's 'updom' to a new reindex domain is incorrect
because the original updom could be unstrided but the new indices may be
strided.

dsiSetIndices should be using 'dist.downdom' to convert the indices
because the current downdom may be empty (e.g., if we just created a new
domain and are initializing it).

chpl_reindexConvertDom should perform a slice between the downdom
dimension and 'actual' range to preserve stride information.

dsiLocalSubdomain now uses chpl_reindexConvertDom to translate the
downward local subdomain to the upward indices.