Fix testTime test for PrgEnv-cray
The type of the field `tm_zone` in `struct tm` is platform dependent. On
darwin it is `char *` and on other platforms it is `const char *`. Set the
type based on CHPL_TARGET_PLATFORM.

Prior to this, the type was set to `c_ptr(c_char)` to match the darwin type.
Now it is `c_ptr(c_char)` for darwin, and `c_string` elsewhere.

Tested on the modules/standard/datetime tests on:
* darwin + clang
* linux64 + gcc
* Cray XC + PrgEnv-cray