Update this test for the different requirements in an initializer world
A type that defines a constructor is also expected to have a default constructor
with an argument per field sharing its name with that field.  A type that
defines an initializer, on the other hand, is not expected to have the compiler
generate a default initializer.  Update this test to reflect that, which allows
it to pass and have its .future file removed.