Update references in QUICKSTART.rst
References point to correct path and use relative html links

2 Problems:
    1. Relative html links don't work on github rendered-version
    2. The updated paths are a bit verbose.

For 1, We could use rst-links instead, which act as no-op insteaad of dead
link on github.

For 2, we could drop references to file-paths and just reference the
page by it's title, e.g. "See the Chapel Environment page for more
details."