single element tuple case for String join() (#5898)
[reviewed by @ben-albrecht and @dmk42]

If a tuple has a single string in it and join() is called with it, join() will attempt to look for a .domain in the tuple, which doesn't exist. This PR checks if the single element collection passed is an array or a tuple before returning, and adds tests for single element tuple and varargs cases.