Use Schubert numbering instead of conditionals for subclass check
Fixes #5887. The subclass check is used in dynamic casting.

See that issue for a description of Schubert numbering
(or see "Implementing statically typed object-oriented programming
languages", Roland Ducournau).

Basically, the strategy of Schubert numbering consists of two components:

1) compute class IDs in a preorder depth-first traversal.
    That way, subclasses of any class form a contiguous region in the
    space of class IDs.

 2) create an array storing for each class the maximum class ID of it or
    a subclass. Since property 1 made the class IDs of all subclasses
    contiguous in any list, this maximum can be used to check if one
    class ID represents a subclass of another. It just amounts to
    two comparisons.

I chose to implement Schubert numbering for the dynamic subclass check
because itis reasonably convenient to implement in the current compiler.
There are certainly other ways to implement a subclass check.